Move shared color helpers into color.rs
Review response: relative_luminance is already provided by the color
crate (OpaqueColor::relative_luminance), and border.rs already had the
same WCAG contrast_ratio. contrast_ratio moves to color.rs (shared with
border.rs) and blend_for_contrast joins it, built on the color crate's
luminance and lerp_rect instead of hand-rolled equivalents.
